Trait substrate_frame_rpc_system::SystemApi

pub trait SystemApi<AccountId, Index>: Sized + Send + Sync + 'static {
    fn nonce(&self, account: AccountId) -> FutureResult<Index>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        AccountId: Send + Sync + 'static + DeserializeOwned,
        Index: Send + Sync + 'static + Serialize,
    { ... }
}

System RPC methods.

Required methods

fn nonce(&self, account: AccountId) -> FutureResult<Index>

Returns the next valid index (aka nonce) for given account.

This method takes into consideration all pending transactions currently in the pool and if no transactions are found in the pool it fallbacks to query the index from the runtime (aka. state nonce).

Loading content...
Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
AccountId: Send + Sync + 'static + DeserializeOwned,
Index: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.

Loading content...