Trait substrate_frame_rpc_system::SystemApi

pub trait SystemApi<BlockHash, AccountId, Index>: Sized + Send + Sync + 'static {
    fn nonce(&self, account: AccountId) -> BoxFuture<Result<Index, RpcError>>;

    fn dry_run(
        &self, 
        extrinsic: Bytes, 
        at: Option<BlockHash>
    ) -> BoxFuture<Result<Bytes, RpcError>>;

    fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M>
    where
        BlockHash: Send + Sync + 'static + DeserializeOwned,
        AccountId: Send + Sync + 'static + DeserializeOwned,
        Index: Send + Sync + 'static + Serialize,
    { ... }
}

System RPC methods.

Required methods

fn nonce(&self, account: AccountId) -> BoxFuture<Result<Index, RpcError>>

Returns the next valid index (aka nonce) for given account.

This method takes into consideration all pending transactions currently in the pool and if no transactions are found in the pool it fallbacks to query the index from the runtime (aka. state nonce).

fn dry_run(
&self,
extrinsic: Bytes,
at: Option<BlockHash>
) -> BoxFuture<Result<Bytes, RpcError>>

Dry run an extrinsic at a given block. Return SCALE encoded ApplyExtrinsicResult.

Provided methods

fn to_delegate<M: Metadata>(self) -> IoDelegate<Self, M> where
BlockHash: Send + Sync + 'static + DeserializeOwned,
AccountId: Send + Sync + 'static + DeserializeOwned,
Index: Send + Sync + 'static + Serialize,

Create an IoDelegate, wiring rpc calls to the trait methods.