Trait substrate_frame_rpc_system::SystemApiClient

pub trait SystemApiClient<BlockHash, AccountId, Index>: ClientTwhere
    BlockHash: Send + Sync + 'static + Serialize,
    AccountId: Send + Sync + 'static + Serialize,
    Index: Send + Sync + 'static + DeserializeOwned,{
    fn nonce<'life0, 'async_trait>(
        &'life0 self,
        account: AccountId
    ) -> Pin<Box<dyn Future<Output = RpcResult<Index>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
    fn dry_run<'life0, 'async_trait>(
        &'life0 self,
        extrinsic: Bytes,
        at: Option<BlockHash>
    ) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>
    where
        Self: Sync + 'async_trait,
        'life0: 'async_trait,
    { ... }
}

Client implementation for the SystemApi RPC API.

Provided Methods

fn nonce<'life0, 'async_trait>(
&'life0 self,
account: AccountId
) -> Pin<Box<dyn Future<Output = RpcResult<Index>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Returns the next valid index (aka nonce) for given account.

This method takes into consideration all pending transactions currently in the pool and if no transactions are found in the pool it fallbacks to query the index from the runtime (aka. state nonce).

fn dry_run<'life0, 'async_trait>(
&'life0 self,
extrinsic: Bytes,
at: Option<BlockHash>
) -> Pin<Box<dyn Future<Output = RpcResult<Bytes>> + Send + 'async_trait>>where
Self: Sync + 'async_trait,
'life0: 'async_trait,

Dry run an extrinsic at a given block. Return SCALE encoded ApplyExtrinsicResult.