substrate_frame_rpc_system

Trait SystemApiClient

pub trait SystemApiClient<BlockHash, AccountId, Nonce>: ClientT
where
    BlockHash: Send + Sync + 'static + Serialize,
    AccountId: Send + Sync + 'static + Serialize,
    Nonce: Send + Sync + 'static + DeserializeOwned,
{
    // Provided methods
    fn nonce<'life0, 'async_trait>(
        &'life0 self,
        account: AccountId,
    ) -> Pin<Box<dyn Future<Output = Result<Nonce, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
    fn dry_run<'life0, 'async_trait>(
        &'life0 self,
        extrinsic: Bytes,
        at: Option<BlockHash>,
    ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
       where Self: Sync + 'async_trait,
             'life0: 'async_trait { ... }
}

Client implementation for the SystemApi RPC API.

Provided Methods

fn nonce<'life0, 'async_trait>( &'life0 self, account: AccountId, ) -> Pin<Box<dyn Future<Output = Result<Nonce, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Returns the next valid index (aka nonce) for given account.

This method takes into consideration all pending transactions currently in the pool and if no transactions are found in the pool it fallbacks to query the index from the runtime (aka. state nonce).

fn dry_run<'life0, 'async_trait>( &'life0 self, extrinsic: Bytes, at: Option<BlockHash>, ) -> Pin<Box<dyn Future<Output = Result<Bytes, Error>> + Send + 'async_trait>>
where Self: Sync + 'async_trait, 'life0: 'async_trait,

Dry run an extrinsic at a given block. Return SCALE encoded ApplyExtrinsicResult.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.